feat: Enhance webhook handling for webhookPlaybackStop and refactor AnilistScrobbler
- Split AnilistScrobbler into a separate file for better organization and maintainability.
- Improved overall readability and maintainability of the playback stop handling.
- Enhanced webhook action to dynamically initialize JellyfinMiniApi based on the server URL.
- Introduced NOTIFICATION_TYPES constant for better notification type management.
- Refactored webhookPlaybackStop to accept JellyfinMiniApi and AnilistScrobbler instances as parameters.
